Move partask and mpi stage to hpcSPAdes
