+++ src/projects/hpcspades/common/CMakeLists.txt
@@ -10,3 +10,4 @@ project(common_modules_mpi CXX)
 add_subdirectory(paired_info)
 add_subdirectory(alignment)
 add_subdirectory(stages)
+add_subdirectory(pipeline)
