+++ src/projects/hpcspades/pair_info_count_mpi.hpp
@@ -9,7 +9,7 @@
 
 #include "projects/spades/pair_info_count.hpp"
 #include "common/alignment/sequence_mapper_notifier_mpi.hpp"
-#include "pipeline/mpi_stage.hpp"
+#include "common/pipeline/mpi_stage.hpp"
 
 namespace debruijn_graph {
     class PairInfoCountMPI : public PairInfoCountBase, public spades::MPIAssemblyStage {
